drop sync daemon, update env config for email, test ocp
