+++ .env.template
@@ -24,11 +24,20 @@ OPENWEATHERMAP_API_KEY=your_openweathermap_key_here
 # Sign up: https://www.mailjet.com/ — free tier sufficient for low-volume alerts
 MAILJET_API_KEY=your_mailjet_api_key_here
 MAILJET_API_SECRET=your_mailjet_api_secret_here
+SMTP_FROM=your_smtp_from_here
+SMTP_TO=your_smtp_to_here,your_smtp_to_here_2
 
 # Storage compaction bucket names (required when storage.compaction.backend is s3, gcs, or abfs in config.toml)
+### S3 Compatible Storage
 S3_BUCKET=
-GCS_BUCKET=
-AZURE_BUCKET=
+S3_ENDPOINT_URL=  # Optional: S3-compatible endpoint URL (e.g. MinIO, DigitalOcean Spaces)
+AWS_ACCESS_KEY_ID=
+AWS_SECRET_ACCESS_KEY=
 
-# Optional: S3-compatible endpoint URL (e.g. MinIO, DigitalOcean Spaces)
-# S3_ENDPOINT_URL=
+### GCS Storage
+GCS_BUCKET=
+GOOGLE_APPLICATION_CREDENTIALS=path/to/your/gcs/credentials.json
+
+### Azure Storage
+AZURE_BUCKET=
+# credentials in line with adlfs
