+++ config.toml
@@ -157,5 +157,7 @@ cooldown_hours = 4  # Minimum hours between repeat alerts
 last_alert_state = ".daemon_state/last_alert.json"  # Cooldown state file
 # API-based provider; only "mailjet" is implemented (MAILJET_API_KEY, MAILJET_API_SECRET in .env)
 email_provider = "mailjet"
+
+# Emails from/to controlled by smtp_from and smtp_to in .env
 smtp_from = "alerts@example.com"  # Sender (must be verified in Mailjet)
 smtp_to = ["oncall@example.com"]  # Recipients
